requirements.txt
pyinstaller
websockets
PyQt5
mss
opencv-python
numpy
Pillow
pynput